Added Utils Models package to dash package. Relocated TimelineConverter to Utils package Moved DashManifestModel to models/ Renamed and Relocated FragmentExtensions to FragmentedTextBoxParser and moved into utils.
