+++ src/dash/DashMetrics.js
@@ -31,7 +31,7 @@
 import HTTPRequest from '../streaming/vo/metrics/HTTPRequest.js';
 import AbrController from '../streaming/controllers/AbrController.js';
 import ManifestModel from '../streaming/models/ManifestModel.js';
-import DashManifestModel from './DashManifestModel.js';
+import DashManifestModel from './models/DashManifestModel.js';
 import FactoryMaker from '../core/FactoryMaker.js';
 
 function DashMetrics() {
